+++ src/main/java/com/example/amigoscode/repository/StudentRepository.java
@@ -13,7 +13,7 @@ import java.util.Optional;
  */
 @Repository
 public interface StudentRepository
-        extends JpaRepository<Student, Long> {
+        extends JpaRepository<Student, Integer> {
 
     Optional<Student> findStudentByEmail(String email);
 }
